+++ src/django_components/__init__.py
@@ -16,7 +16,7 @@ from django_components.util.command import (
     CommandSubcommand,
     ComponentCommand,
 )
-from django_components.component import Component, ComponentVars, all_components
+from django_components.component import Component, ComponentVars, all_components, get_component_by_class_id
 from django_components.component_media import ComponentMediaInput, ComponentMediaInputPath
 from django_components.component_registry import (
     AlreadyRegistered,
@@ -96,6 +96,7 @@ __all__ = [
     "EmptyTuple",
     "EmptyDict",
     "format_attributes",
+    "get_component_by_class_id",
     "get_component_dirs",
     "get_component_files",
     "get_component_url",